Switch more feature specs to poltergeist
These ones were using Capybara-webkit. All else equal, it's better to run these in poltergeist so that tests will fail when there are javascript errors. Ideally, we should get to the point where we remove Capybara-webkit altogether, and maintain a set of tests that run correctly with either Poltergeist or Selenium, simply by toggling the current_driver.